Create bootstrap/workflow for integration testing
You can now write and run integration tests for the server using the
bootstrap that is provided in `test/lib/bootstrap.js`.

Take a look at `tests/integration/api.js` for examples of how to write a
test.

You can run the tests with gulp (make sure to `npm install -g gulp`)
with `gulp` or `gulp mocha`.

I've included a README with instructions.